🔨 Refactor: vemo page 캡처하기 그리기 등 refactor
